refactor: inference tests get own directory and shared helper module
